+++ crates/gen/src/types/mod.rs
@@ -4,7 +4,6 @@ mod bool32;
 mod bstr;
 mod class;
 mod com_interface;
-mod delegate;
 mod r#enum;
 mod handle;
 mod interface;
@@ -23,7 +22,6 @@ pub use bool32::*;
 pub use bstr::*;
 pub use class::*;
 pub use com_interface::*;
-pub use delegate::*;
 pub use handle::*;
 pub use interface::*;
 pub use matrix3x2::*;
